+++ tests/thread.c
@@ -10,7 +10,7 @@ void *thread_fn(void* arg) {
   l = 4;
   for (volatile int i = 0; i < N; i++) {
     if ((i % (N/10)) == 0) {
-      printf("------- Hello Thread | Self: %d, TL: %d, K: %d\n", pthread_self(), l, k++);
+      printf("------- Hello Thread | Self: %d, TL: %d, K: %d\n", (int)pthread_self(), l, k++);
       //PRINT_INT("---------- Hello Thread | Self", (long)pthread_self());
     }
   }
@@ -25,7 +25,7 @@ int main() {
   int k = 0;
   for (volatile int i = 0; i < N*2; i++) {
     if ((i % (N/10)) == 0) {
-      printf("------ Hello Parent | Self: %d, TL: %d, K: %d\n", pthread_self(), l, k++);
+      printf("------ Hello Parent | Self: %d, TL: %d, K: %d\n", (int)pthread_self(), l, k++);
     }
   }
   return 0;
